Added a difficulty setting and endpoint to get it There is not a 'Difficulty' enum with 4 distinct settings: 	- Easy 	- Medium 	- Hard 	- Expert
There is a corresponding value in the 'Player' class to store
the difficulty setting for that player.

The new server endpoint '/difficulty' can be accessed like this:

http://localhost:3000/difficulty?id=ID

Where ID is the id of the player whos difficulty level you would
like to obtain. It returns a string corresponding to the level
as well as status 200 on success, 404 if the player does not exist
or is offline, and 400 if ID is not a number.